Rename dg_object.c
